+++ ci/test_all_features.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
+
+std=''
+if [ "${1:-}" = 'std' ]; then
+    std=',std'
+fi
+
 set -euxo pipefail
 
-for feature in $(tomljson Cargo.toml | jq --raw-output '.features | keys[]' | grep -v 'default\|std\|testing-helpers'); do
-    cargo test -p derive_more --tests --no-default-features --features "$feature,testing-helpers";
+for feature in $(tomljson Cargo.toml | jq --raw-output '.features | keys[]' | grep -v 'default\|std\|full\|testing-helpers'); do
+    cargo +nightly test -p derive_more --tests --no-default-features --features "$feature$std,testing-helpers"
 done